re-arrange the structure
