+++ modules/face/samples/facemark_demo_aam.cpp
@@ -39,6 +39,7 @@
 
 using namespace std;
 using namespace cv;
+using namespace cv::face;
 
 Mat loadCSV(std::string filename);
