+++ modules/face/samples/facemark_demo_lbf.cpp
@@ -31,6 +31,7 @@
 
  using namespace std;
  using namespace cv;
+using namespace cv::face;
 
  int main(int argc, char** argv )
  {
